refactor: move pl::Ty(Kind) Display impls to pl module
The current "codegen" module implements source code formatting.
A large portion of the source code currently relies on Expr
implementing Display. Since we want to move Expr to a new prql_ast
crate we for now also have to move the Display impl there (since
you cannot impl Display for a foreign type), and since the Display
impl currently relies on the "codegen" module, we'll have to move
it along as well.

However we certainly don't want the type for types (Ty) in prql_ast,
so we firstly move the Display impls for Ty & TyKind, which should stay
here to another module. Doing this in a separate commit helps the git
file rename detection.
